refactor di: merge login and authhostservice components
